requirements.txt
openai>=1.30.0
python-dotenv>=1.0.0
streamlit>=1.38.0

# ML router
scikit-learn>=1.3.0
joblib>=1.3.0

# Tests
pytest>=7.0.0
